migrate mock class to a mock_classes.py package for better sectionability
