+++ utilities/segmentation_utils/reading_strategies.py
@@ -1,37 +1,12 @@
 import os
 from types import ModuleType
-from typing import Protocol, Type, Union
+from typing import Protocol, Union
 
 import numpy as np
 import rasterio
 from PIL import Image
 
-
-class MockRasterio:
-    def __init__(self, n, size, bands, dtypes):
-        self.n = n
-        self.size = size
-        self.bands = bands
-        self.dtypes = dtypes
-
-    def open(self, *args, **kwargs):
-        return self
-
-    @property
-    def count(self) -> int:
-        return self.bands
-
-    def read(self, *args, **kwargs):
-        return np.zeros((self.bands, self.size[0], self.size[1]), self.dtypes[0])
-
-    # these functions are invoked when a 'with' statement is executed
-    def __enter__(self):
-        # called at the beginning of a 'with' block
-        return self  # returns instance of MockRasterio class itself
-
-    def __exit__(self, type, value, traceback):
-        # called at the end of a 'with' block
-        pass
+from tests.mock_classes import MockRasterio
 
 
 class IReader(Protocol):
@@ -64,7 +39,6 @@ class RGBImageStrategy:
         images = np.zeros((batch_size, self.image_size[0], self.image_size[1], 3))
 
         for i in range(batch_size):
-            image_index = i + dataset_index
             image = Image.open(
                 os.path.join(self.image_path, batch_filenames[i])
             ).resize(self.image_size, self.image_resample)
@@ -95,7 +69,6 @@ class HyperspectralImageStrategy:
         self.bands = package.open(
             os.path.join(self.image_path, self.image_filenames[0])
         ).count
-        print("-----------My very cool bands--------: ", self.bands)
 
     def read_batch(self, batch_size: int, dataset_index: int) -> np.ndarray:
         # read images with rasterio
